Ryan

+++ db.js
@@ -1,21 +1,21 @@
 const { Sequelize } = require('sequelize')
 
-const sequelize = new Sequelize(
-    process.env.DB_DBNAME,
-    process.env.DB_USER,
-    process.env.DB_PASS,
+const sequelize = new Sequelize( process.env.DATABASE_URL,
+    // process.env.DB_DBNAME,
+    // process.env.DB_USER,
+    // process.env.DB_PASS,
     {
-        host: process.env.DB_HOST,
+        //host: process.env.DB_HOST,
         dialect: 'postgres'
     }
 )
 
-async function synceDb(sequelize, force = false){
+async function synceDb(sequelize, options){
+    const {force, alter } = options
     try {
-        if (force)
-            await sequelize.sync({force: true})
-        else
-            await sequelize.sync()
+        if (force) await sequelize.sync({force: true})
+        else if (alter) await sequelize.sync({ alter: true})
+        else await sequelize.sync()
     } catch (err){
         console.log(err)
     }
